requirements.txt
torch>=1.7.0
torchvision>=0.8.1
Pillow>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
prometheus-client>=0.12.0
pydantic>=1.8.2
dvc>=2.0.0
dvclive>=0.7.0
numpy>=1.19.4
evidently==0.6.7
schedule==1.2.0 